+++ pocketc_backend/src/main/java/com/ssafy/pocketc_backend/domain/streak/entity/Streak.java
@@ -23,7 +23,7 @@ public class Streak {
     private User user;
 
     private Integer missionCompletedCount;
-    private Integer spentAmount;
+    private Long spentAmount;
 
     @Builder.Default
     private boolean status = false;
